重要版本节点。1.修改日志格式，添加时间记录，精确到秒。2.修改初始化相机时打开相机以及开始采集图像数据并设置关联的HWND。3.修改图像路径导致的识别模块和人工输入模块带来的bug。4.去掉所有的网络WSAStartup和WSACleanup，改为，在整个应用程序初始化的时候调用一次afxSocketinit。

+++ CarSeat_recognization/CarSeat_Recognization/common/RecogResultManager.cpp
@@ -65,7 +65,7 @@ void CRecogResultManager::init()
 			break;
 		}
 		memset(tmpChar, 0, sizeof(char) * 3 * MAX_CHAR_LENGTH);
-		int i = end - content - begin;
+		int64_t i = end - content - begin;
 		memcpy(tmpChar, content + begin, i);
 		if (tmpChar[i - 1] <= 0x20)
 		{
